fix: add readonly to promise type for flexibility
When using this library with `algoliasearch` (whose return types are
always `Readonly<Promise<T>>`), I've had to use a wrapper function so
that I can update those types.

This allows for more flexibility (i.e. by reducing the required
properties on the `Promise` object) and for use with more restrictive
libraries like `algoliasearch`.